// src/messages/types/MessagesEnum.ts
export const enum MessagesEnum {
  WELCOME = 'WELCOME',
  SUBSCRIBE = 'SUBSCRIBE',
  HELP = 'HELP',
  SUBSCRIPTION_NOT_FOUND = 'SUBSCRIPTION_NOT_FOUND',
  ACCESS_DENIED = 'ACCESS_DENIED',
  IN_PROGRESS = 'IN_PROGRESS',
  ERROR = 'ERROR',
  SUBSCRIPTION_DELETED = 'SUBSCRIPTION_DELETED',
  WRONG_DATE_FORMAT = 'WRONG_DATE_FORMAT',
  SUBSCRIPTION_ADDING_ERROR = 'SUBSCRIPTION_ADDING_ERROR',
}